viewPager correção layout
