--- app/src/main/res/layout/fragment_view_pager.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    tools:context=".ui.fragment.ViewPagerFragment">
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/container"
-        android:paddingStart="35dp"
-        android:paddingEnd="35dp"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
